+++ package.json
@@ -88,7 +88,7 @@
 		"prettier": "^3.3.3",
 		"tsup": "^8.3.0",
 		"typescript": "^5.6.2",
-		"vitest": "^2.1.1"
+		"vitest": "^3.0.5"
 	},
 	"keywords": [
 		"ORM",
